+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/DeedOfNovationAndVariationTask/DeedOfNovationAndVariationTask.cshtml.cs
@@ -2,8 +2,6 @@ using Dfe.Complete.Application.Projects.Commands.TaskData;
 using Dfe.Complete.Constants;
 using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Domain.ValueObjects;
-using Dfe.Complete.Extensions;
-using Dfe.Complete.Models;
 using Dfe.Complete.Pages.Projects.TaskList.Tasks.HandoverWithDeliveryOfficerTask;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
@@ -54,7 +52,7 @@ namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.DeedOfNovationAndVariationT
         {
             await sender.Send(new UpdateDeedOfNovationAndVariationTaskCommand(
                 new TaskDataId(TasksDataId.GetValueOrDefault())!, Received, Cleared, SignedOutgoingTrust, SignedIncomingTrust, Saved, SignedSecretaryState, SaveAfterSign));
-            TempData.SetNotification(NotificationType.Success, "Success", "Task updated successfully");
+            SetTaskSuccessNotification();
             return Redirect(string.Format(RouteConstants.ProjectTaskList, ProjectId));
         }
     }